+++ plotly/src/common/color.rs
@@ -25,7 +25,7 @@ pub trait Color {
     fn to_color(&self) -> ColorWrapper;
 }
 
-#[derive(Debug)]
+#[derive(Debug, Clone, Copy)]
 pub struct Rgb {
     r: u8,
     g: u8,
@@ -44,7 +44,7 @@ impl Color for Rgb {
     }
 }
 
-#[derive(Debug)]
+#[derive(Debug, Clone, Copy)]
 pub struct Rgba {
     r: u8,
     g: u8,
@@ -68,7 +68,7 @@ impl Color for Rgba {
 }
 
 // https://www.w3.org/TR/css-color-3/#svg-color
-#[derive(Debug)]
+#[derive(Debug, Clone, Copy)]
 pub enum NamedColor {
     AliceBlue,
     AntiqueWhite,